Moved some more minor functions from Eagle.ts out into other files.
